HDXDSYS-1043 add returnees
