--- src/hapi/pipelines/configs/refugees.yaml
@@ -1,78 +1,0 @@
-refugees_national:
-  refugees:
-    dataset: "unhcr-population-data-for-world"
-    resource: "Demographics and locations of forcibly displaced and stateless persons (Global)"
-    format: "csv"
-    headers: 1
-    use_hxl: True
-    admin:
-      - "#country+code+origin"
-    admin_exact: True
-    input:
-      - "#date+year"
-      - "#country+code+asylum"
-      - "#indicator+population_type"
-      - "#affected+f+infants+age_0_4"
-      - "#affected+f+children+age_5_11"
-      - "#affected+f+adolescents+age_12_17"
-      - "#affected+f+adults+age_18_59"
-      - "#affected+f+elderly+age_60"
-      - "#affected+f+total"
-      - "#affected+m+infants+age_0_4"
-      - "#affected+m+children+age_5_11"
-      - "#affected+m+adolescents+age_12_17"
-      - "#affected+m+adults+age_18_59"
-      - "#affected+m+elderly+age_60"
-      - "#affected+m+total"
-      - "#affected+all+total"
-    list:
-      - "#date+year"
-      - "#country+code+asylum"
-      - "#indicator+population_type"
-      - "#affected+f+infants+age_0_4"
-      - "#affected+f+children+age_5_11"
-      - "#affected+f+adolescents+age_12_17"
-      - "#affected+f+adults+age_18_59"
-      - "#affected+f+elderly+age_60"
-      - "#affected+f+total"
-      - "#affected+m+infants+age_0_4"
-      - "#affected+m+children+age_5_11"
-      - "#affected+m+adolescents+age_12_17"
-      - "#affected+m+adults+age_18_59"
-      - "#affected+m+elderly+age_60"
-      - "#affected+m+total"
-      - "#affected+all+total"
-    output:
-      - "#date+year"
-      - "#country+code+asylum"
-      - "#indicator+population_type"
-      - "#affected+f+infants+age_0_4"
-      - "#affected+f+children+age_5_11"
-      - "#affected+f+adolescents+age_12_17"
-      - "#affected+f+adults+age_18_59"
-      - "#affected+f+elderly+age_60plus"
-      - "#affected+f+total"
-      - "#affected+m+infants+age_0_4"
-      - "#affected+m+children+age_5_11"
-      - "#affected+m+adolescents+age_12_17"
-      - "#affected+m+adults+age_18_59"
-      - "#affected+m+elderly+age_60plus"
-      - "#affected+m+total"
-      - "#affected+all+total"
-    output_hxl:
-      - "#date+year"
-      - "#country+code+asylum"
-      - "#indicator+population_type"
-      - "#affected+f+infants+age_0_4"
-      - "#affected+f+children+age_5_11"
-      - "#affected+f+adolescents+age_12_17"
-      - "#affected+f+adults+age_18_59"
-      - "#affected+f+elderly+age_60plus"
-      - "#affected+f+total"
-      - "#affected+m+infants+age_0_4"
-      - "#affected+m+children+age_5_11"
-      - "#affected+m+adolescents+age_12_17"
-      - "#affected+m+adults+age_18_59"
-      - "#affected+m+elderly+age_60plus"
-      - "#affected+m+total"
-      - "#affected+all+total"
