+++ tests/test_main.py
@@ -21,6 +21,7 @@ from hapi_schema.db_population import DBPopulation
 from hapi_schema.db_poverty_rate import DBPovertyRate
 from hapi_schema.db_refugees import DBRefugees
 from hapi_schema.db_resource import DBResource
+from hapi_schema.db_returnees import DBReturnees
 from hapi_schema.db_sector import DBSector
 from hapi_schema.db_wfp_commodity import DBWFPCommodity
 from hapi_schema.db_wfp_market import DBWFPMarket
@@ -57,7 +58,7 @@ class TestHAPIPipelines:
             "operational_presence.yaml",
             "population.yaml",
             "poverty_rate.yaml",
-            "refugees.yaml",
+            "refugees_and_returnees.yaml",
             "wfp.yaml",
         ]
         project_config_dict = load_yamls(project_configs)
@@ -198,15 +199,19 @@ class TestHAPIPipelines:
         )
         check.equal(count, 25)
 
-    @pytest.mark.parametrize("themes_to_run", [{"refugees": None}])
-    def test_refugees(self, configuration, folder, pipelines):
+    @pytest.mark.parametrize(
+        "themes_to_run", [{"refugees_and_returnees": None}]
+    )
+    def test_refugees_and_returnees(self, configuration, folder, pipelines):
         session = pipelines.session
         count = session.scalar(select(func.count(DBDataset.hdx_id)))
         check.equal(count, 1)
         count = session.scalar(select(func.count(DBResource.hdx_id)))
         check.equal(count, 1)
         count = session.scalar(select(func.count(DBRefugees.resource_hdx_id)))
-        check.equal(count, 102726)
+        check.equal(count, 570921)
+        count = session.scalar(select(func.count(DBReturnees.resource_hdx_id)))
+        check.equal(count, 5408)
 
     @pytest.mark.parametrize("themes_to_run", [{"idps": None}])
     def test_idps(self, configuration, folder, pipelines):
